requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
pytest==8.0.0
httpx==0.26.0
python-dotenv==1.0.1
email-validator==2.1.0.post1 